Day 4/requirements.txt
supabase
python-docx
python-dotenv
google-genai
pydantic
fastapi
uvicorn
